# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Remedy Scribe API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/audit-logs": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "List Audit Logs",
        "description": "The drill-down: every matching audit row, newest first.\n\n`action_prefix` exists because the vocabulary is deliberately dotted\n(see app/services/audit.py) \u2014 \"show me everything anyone did to notes\"\nis `note.`, and \"every disclosure of audio\" is\n`encounter.audio.`. Without it a reviewer has to know the exact action\nstrings before they can look for anything.",
        "operationId": "list_audit_logs_api_v1_audit_logs_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "actor_clinician_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Everything this clinician did",
              "title": "Actor Clinician Id"
            },
            "description": "Everything this clinician did"
          },
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "e.g. \"patient\", \"note\", \"encounter\"",
              "title": "Entity Type"
            },
            "description": "e.g. \"patient\", \"note\", \"encounter\""
          },
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Everything done to this one record",
              "title": "Entity Id"
            },
            "description": "Everything done to this one record"
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact action, e.g. note.grounding.read",
              "title": "Action"
            },
            "description": "Exact action, e.g. note.grounding.read"
          },
          {
            "name": "action_prefix",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Dotted prefix, e.g. \"note.\" or \"encounter.upload.\"",
              "title": "Action Prefix"
            },
            "description": "Dotted prefix, e.g. \"note.\" or \"encounter.upload.\""
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive lower bound on created_at",
              "title": "Since"
            },
            "description": "Inclusive lower bound on created_at"
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exclusive upper bound on created_at",
              "title": "Until"
            },
            "description": "Exclusive upper bound on created_at"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogEntryOut"
                  },
                  "title": "Response List Audit Logs Api V1 Audit Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit-logs/access-report": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Access Report",
        "description": "**\"Who looked at this record, and when?\"** \u2014 the question P0-8 exists\nto make answerable, in one call.\n\nOne row per (actor, action), ordered by most recent activity, with the\nactor's real name resolved so the report reads as names rather than\nUUIDs. `LEFT OUTER JOIN` on the clinician: `actor_clinician_id` is\nnullable (a system-initiated action has no human actor) and the row\nmust still appear \u2014 an unattributed access is the *most* interesting\nline in a breach investigation, and an inner join would silently drop\nexactly those.\n\nNote that this only reports what the trail contains. A record with no\nrows returns an empty report, which means \"no logged access\", not \"no\naccess\" \u2014 the two are the same claim only because every disclosure path\nin the API writes here (Phase 4.2's actual work).",
        "operationId": "access_report_api_v1_audit_logs_access_report_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entity_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The record's type, e.g. \"patient\", \"note\", \"encounter\"",
              "title": "Entity Type"
            },
            "description": "The record's type, e.g. \"patient\", \"note\", \"encounter\""
          },
          {
            "name": "entity_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The record's id",
              "title": "Entity Id"
            },
            "description": "The record's id"
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive lower bound on created_at",
              "title": "Since"
            },
            "description": "Inclusive lower bound on created_at"
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exclusive upper bound on created_at",
              "title": "Until"
            },
            "description": "Exclusive upper bound on created_at"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditAccessReportRow"
                  },
                  "title": "Response Access Report Api V1 Audit Logs Access Report Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditAccessReportRow": {
        "properties": {
          "actor_clinician_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Clinician Id"
          },
          "actor_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Email"
          },
          "actor_full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Full Name"
          },
          "actor_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Role"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "access_count": {
            "type": "integer",
            "title": "Access Count"
          },
          "first_at": {
            "type": "string",
            "format": "date-time",
            "title": "First At"
          },
          "last_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last At"
          }
        },
        "type": "object",
        "required": [
          "actor_clinician_id",
          "actor_email",
          "actor_full_name",
          "actor_role",
          "action",
          "access_count",
          "first_at",
          "last_at"
        ],
        "title": "AuditAccessReportRow",
        "description": "One (actor, action) pair in the access report for a single record.\n\nGrouped rather than listed because the question P0-8 exists to answer \u2014\n\"who looked at this patient's record?\" \u2014 is answered by a handful of\nnames, not by five hundred individual rows. The raw rows are still one\n`GET /audit-logs?entity_id=...` away.\n\nThe actor's name/email/role are joined in here (they are staff\nidentity, not PHI) because an audit report naming only UUIDs is one\nnobody can review, which is the failure mode \"reviewable\" is guarding\nagainst. They are read live from `clinicians`, never copied into\n`audit_logs`, so a renamed or deactivated account reports its current\nidentity and a deleted one degrades to nulls rather than to a stale\nname frozen years ago."
      },
      "AuditLogEntryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "actor_clinician_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Clinician Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "diff": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diff"
          },
          "retention_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Retention Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "actor_clinician_id",
          "action",
          "entity_type",
          "entity_id",
          "created_at",
          "diff",
          "retention_expires_at"
        ],
        "title": "AuditLogEntryOut",
        "description": "One audit row as the review interface renders it.\n\n`diff` is safe to expose because nothing may put PHI in it \u2014 see\n`app/models/audit_log.py`'s class docstring. `retention_expires_at` is\nexposed because a reviewer looking at a nearly-expired row needs to\nknow it is about to become unavailable, and because it is the value the\nappend-only trigger keys off: before this date the row cannot be\ndeleted by anyone, including the app's own DB role."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}